requirements.txt
beautifulsoup4>=4.12
lxml>=5.0
tqdm>=4.66
sentence-transformers>=2.7
chromadb>=0.5.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.3.0
fastapi>=0.115
uvicorn>=0.30
httpx>=0.27
pyyaml>=6.0
